0.2.0
added script commands to package.json, updatedserver/server.js to console log the url so that it is clickable inside the console. updated client/js/install.js to handle the display of the prompts.  updated webpack.config.js so that the correct entry point and output bundles are provided, injects the service worker, creates a manifest.json. and correctly loads stylesheets etc.